+++ src/ODBC2/sqlite/pager.c
@@ -452,7 +452,7 @@ static int readMasterJournal(OsFile *pJrnl, char **pzMaster){
   }
 
   /* See if the checksum matches the master journal name */
-  for(i=0; i<len; i++){
+  for(i=0; i< (int) len; i++){
     cksum -= (*pzMaster)[i];
   }
   if( cksum ){
@@ -947,12 +947,12 @@ static int pager_delmaster(const char *zMaster){
     /* Load the entire master journal file into space obtained from
     ** sqliteMalloc() and pointed to by zMasterJournal. 
     */
-    zMasterJournal = (char *)sqliteMalloc(nMasterJournal);
+    zMasterJournal = (char *)sqliteMalloc( (int)nMasterJournal);
     if( !zMasterJournal ){
       rc = SQLITE_NOMEM;
       goto delmaster_out;
     }
-    rc = sqlite3OsRead(&master, zMasterJournal, nMasterJournal);
+    rc = (int) sqlite3OsRead(&master, zMasterJournal, (int) nMasterJournal);
     if( rc!=SQLITE_OK ) goto delmaster_out;
 
     zJournal = zMasterJournal;
@@ -1153,7 +1153,7 @@ static int pager_playback(Pager *pPager){
     */
     if( nRec==0xffffffff ){
       assert( pPager->journalOff==JOURNAL_HDR_SZ(pPager) );
-      nRec = (szJ - JOURNAL_HDR_SZ(pPager))/JOURNAL_PG_SZ(pPager);
+      nRec = (u32) ( (szJ - JOURNAL_HDR_SZ(pPager))/JOURNAL_PG_SZ(pPager) );
     }
 
     /* If this is the first header read from the journal, truncate the
@@ -1173,7 +1173,7 @@ static int pager_playback(Pager *pPager){
   
     /* Copy original pages out of the journal and back into the database file.
     */
-    for(i=0; i<nRec; i++){
+    for(i=0; i< (int) nRec; i++){
       rc = pager_playback_one_page(pPager, &pPager->jfd, 1);
       if( rc!=SQLITE_OK ){
         if( rc==SQLITE_DONE ){
@@ -1294,7 +1294,7 @@ static int pager_stmt_playback(Pager *pPager){
     goto end_stmt_playback;
   }
   pPager->journalOff = pPager->stmtJSize;
-  pPager->cksumInit = pPager->stmtCksum;
+  pPager->cksumInit = (u32) pPager->stmtCksum;
   assert( JOURNAL_HDR_SZ(pPager)<(pPager->pageSize+8) );
   while( pPager->journalOff <= (hdrOff-(pPager->pageSize+8)) ){
     rc = pager_playback_one_page(pPager, &pPager->jfd, 1);
@@ -1311,7 +1311,7 @@ static int pager_stmt_playback(Pager *pPager){
       goto end_stmt_playback;
     }
     if( nRec==0 ){
-      nRec = (szJ - pPager->journalOff) / (pPager->pageSize+8);
+      nRec = (u32) ( (szJ - pPager->journalOff) / (pPager->pageSize+8) );
     }
     for(i=nRec-1; i>=0 && pPager->journalOff < szJ; i--){
       rc = pager_playback_one_page(pPager, &pPager->jfd, 1);
@@ -1597,9 +1597,9 @@ int sqlite3pager_pagecount(Pager *pPager){
     n++;
   }
   if( pPager->state!=PAGER_UNLOCK ){
-    pPager->dbSize = n;
+    pPager->dbSize = (int) n;
   }
-  return n;
+  return (int) n;
 }
 
 /*
@@ -1663,7 +1663,7 @@ static void memoryTruncate(Pager *pPager){
 
   ppPg = &pPager->pAll;
   while( (pPg = *ppPg)!=0 ){
-    if( pPg->pgno<=dbSize ){
+    if( (int) pPg->pgno <= dbSize ){
       ppPg = &pPg->pNextAll;
     }else if( pPg->nRef>0 ){
       memset(PGHDR_TO_DATA(pPg), 0, pPager->pageSize);
